Handle TLS in multi-region environments
In a multi-region environment, Keystone is expected to run only
in the primary region (region controller).

At the same time, each region will have its own Vault application,
which can optionally use the same intermediate CAs.

That being considered, we'll update the tls feature to upload the
CAs to the region controller, adding the region name to the CA
names to avoid overlaps.